+++ src/components/Indicator/Indicator.tsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import { ReactComponent as Icon } from './icon.svg';
+
+export const Indicator = () => <Icon />;